Add printBoard method and edit @spaces
This method will traverse through the @spaces array and print the
elements along with separators to resemble a tic tac toe board.

Also edited the default value of the @spaces instance variable to a " "
(space) so that it would not grow in size when it gets filled with a
symbol.