Let Player choose space, Board checks availability
Player class has choose_space method that takes in an X-coordinate and
Y-coordinate. Also added an attribute accessor for the symbol and
coordinates so that they can be accessed by the main program.

Board also check the availability of the space and makes sure it's empty
before adding something to it.