Add checks to user input and out-of-bounds arrays
Added game logic that will allow each user a turn and will stop the game
if either user wins after their turn. Also added logic that prevents a
user from accessing a space already occupied, or a space outside of the
range.

Also changed the wording of UI to make it easier to select area.